Move chooser/ to single/ folder instead
